Move code outside  __init__
